feat: Add in an interactive counter using the DOM interface
The DOM interface allows for dynamic content to be rendered on screen including
user interaction like this button to increment the count number shown.

While this is powerful, this simple set of HTML elements takes multiple
statements to render each item, includes repetition and an imperative code style
which makes re-using this code and state harder to do.