feat: Add support for class components to createElement
While adding the Component class made it easy to encapsulate state, our final
code to render to the page had begun to diverge and repetition had begun to
creep in.

Instead, we update createElement to understand how to use a Component class
correctly to render a DOM element that can be added to the page.

We still haven't tackled nesting of elements for more complex structures.
We do have some nesting in our Counter component, but we've returned to an
imperative code style where several calls are required to nest children into
the root element of the component.