feat: Add support for function components to createElement
One way to deal with nesting is to group createElement calls into a function
which makes the group reusable. It also removes the need for us to call `render`
multiple times at the top level.

To do this, we add support to createElement to understand how to work with
non-class functions in order to produce a DOM element.