feat: Support primitive children
Thus far, child elements have always been DOM elements that can be attached to
the page. Where we've needed to add text to the page, we've used DOM features
like `textContent`. But this approach has draw backs now we have the benefits
of JSX. The syntax to display text content in an element is far more boilerplate
than anything else in the component.

JSX expects any compatible `createElement` function to be able to accept non-object
values in the `children` parameter of the function. So we check for this when
we iterate over this parameter and, if a primitive child is found, we create a
TextNode, which is another native browser type, to display it.