feat: Add a Toggle component
We've now added a component that has two interactions with the page, and our
system is starting to look a little unstable.

Firstly, even though we're using JSX and our createElement helper under the hood,
when we add interaction, we're not actually using any of this. In Toggle, each
interaction directly updates the HTML element directly, and the templated
text is actually duplicated in order to change it each time.

Even worse, if we were needed to call the global render function again in app.jsx
(maybe because some props have changed elsewhere in the app), we would lose both
our Component states which, in a real world scenario, would mean users could
lose data or may have to start a process from scratch.

We need a way to:
- Re-run our Component.render function in order to update to state changes
- Maintain state even when other component need to re-render from scratch