refactor: Introduce virtual DOM elements
As we saw in the previous commit, our previous approach had reached the limits
of its usefulness. While we could've continued to investigate fixes to the
issues it presented, determining the difference between one render and another
would be time consuming and potentially result in unstable or resource hungry
code.

Let's look at the issues we're facing:
- We want to render some HTML and then, when user interaction happens, render
  it again with the changes that interaction made
- Where we have class Components with long-lived state we want to retain that
  state so it persists over multiple interactions
- We want to be able to re-render based on external effects (such as a URL change)
  and have the ability to add, remove and update HTML elements

Since ~2010, we've seen a variety of approaches to solve these problems.
React, alongside many other frameworks, popularised the use of the virtual DOM
as an approach to solving this problem, while also reducing the overhead of
making hundreds of changes directly via the DOM API, which is comparatively
slow for comparing changes.

We're going to mimic this approach by refactoring our `createElement` to
produce a plain object containing the information required to render it and
its children. The returned object from `createElement` will now look like:

{
    // This allows us to hold a reference to a class component
    type: string | Component,
    // By merging props together with children, parent components can interact with children
    props: {
        // These are normal props, e.g. { className: 'heading' } or { onClick: () => {...} }
        [key: string]: string | Function,
        // Merge children nodes in with other properties
        children: Array<VNode | string | number>
    }
}

Note that loading the page with this commit WILL NOT WORK. Instead, if you open
your developer tools console, you'll see each VNode that `createElement` creates
to get an idea of how components are rendered.