refactor: Represent text elements in the virtual DOM
The final change we need to make is to bring createTextElement in line with the
rest of move to vDOM elements. Just as in the DOM API we have a special TextNode
type to represent primitive values, we add a special vNode type value to
represent text only elements. This will allow us to map these values back to
TextNode objects when we convert vDOM elements to DOM elements.