feat: Render a vNode to a real DOM element on the page
To get us back to where we were with our previous DOM implementation, we use
some code that should be fairly familiar. A good deal of this code has been
directly lifted from our old version of `createElement`.

Instead of directly taking the information passed to `createElement`, we take
the vDOM element that `createElement` produced and create all the necessary
DOM elements and TextNodes.

To ensure we can render a class Component, we keep very similar logic to
what we had previously, creating a new instance of the class and immediately
running the render function.

The downside to this is that, because the Component never has direct access
to the DOM, we have lost our interactivity entirely. We need a reconciliation
step in order to return these state updates to a working order.