feat: Introduce vNode comparison
One of the often touted benefits of using a virtual DOM approach is that two
versions of the same tree of elements can be compared to determine what the
differences are between them and what operations are needed on the real DOM
in order to get from version A to version B.

Here we introduce the most naive version of this diffing algorithm. The
`reconcile` function takes a previously rendered and yet to be rendered vNode
and does simple existence checks on both elements to determine if a creation,
removal or replacement operation is needed.

To aid in this, we move our previous render code to a function which will
create a RenderedVNode object, which links a vNode to a DOM element as well
as handling doing the same for a vNode's children.

In its current state, nothing will change. However, the next commit will
demonstrate how this allows us to begin reacting to external changes by
allowing for multiple re-renders.