feat: Demonstrate reconcile diffing
In this slightly contrived offshoot, we render a single element with the timestamp shown.
We render the same component every second, and count the iterations. After 10 iterations,
we pass a `null` value to remove the element from the page.

In doing so, we demonstrate the three scenarios that our naive `reconcile` method caters
for, first adding, then updating, and finally removing the DOM element from the page.

We'll revert the changes made in this commit and continue on to use our new diffing
capabilities to return interactivity to our page.