feat: Add class Component support to reconciliation
We make a considerable change to several parts of both vDOM and DOM libraries to
provide persistent support for class Components.

First, we make changes to `instantiate` to create a component instance and then
attach rendering information to the instance, in order to make updates to just
that part of the page in future updates.

Our Component class now has a `setState` method, which allows state to be updated
and also, if set, calls a `__triggerUpdate` function every time state changes.
This allows our rendering library (in this case 'src/react-dom.js') to run the
diffing algorithm to compare the currently rendered version and a new version created
by the class. In doing so, we're able to show updated versions of the component
as new state changes occur.

Finally, our `reconcile` recognises vNode's with a complex component type and, instead
of starting from scratch, runs a component re-render to maintain the component
instance state.

In the next commit, we'll update our current components to use this new data flow
instead of the current method that no longer works with the vDOM approach.