feat: Update class components to use proper state data flow
Both Counter and Toggle have been updated to use an API almost identical to the one
used by React. The `render` functions of both components are now pure, producing
the exact same output given the same inputs, and changes to the state are made through
the `setState` method on the Component class.

Running the project now should result in a fully interactive page. However, we still
have two problems left which are in scope for this demonstration:
- We have no way of reacting to component mounting or component updates as we would in React
- Because the Toggle component is rendered within the Counter component, the state of
  Toggle is lost if the state of Counter changes. This is because we make no re-use of
  existing DOM elements, even if they are identical between renders

In the next commits, we'll tackle these final problems.